Define `hindent` for the easy use (#709)
* Create 'HIndent.CommandlineOptions'

* Define `hindent`

* Remove unnecessary imports

* Docs
